+++ tests/_data/StubViewProject/src/Service/StdClass.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace StubViewProject\Service;
+
+use Phalcon\Di;
+use Phapp\Application\Service\InjectableInterface;
+
+class StdClass implements InjectableInterface
+{
+    public static function injectTo(Di $di)
+    {
+        $di->setShared(
+            'stdClass',
+            function () use ($di) {
+                return new \stdClass;
+            }
+        );
+    }
+}