+++ tests/functional/_bootstrap.php
@@ -1,2 +1,3 @@
 <?php
 // Here you can initialize variables that will be available to your tests
+define('CODECEPT_DATA_DIR', codecept_data_dir());